requirements.txt
#instagrapi==2.1.5
instagrapi==1.16.37
psycopg2-binary==2.9.7
sqlalchemy==2.0.19
pandas==2.0.3
numpy==1.24.3
python-dotenv==1.0.0
requests==2.31.0
beautifulsoup4==4.12.2
matplotlib==3.7.2
seaborn==0.12.2
plotly==5.15.0
dash==2.11.1
schedule==1.2.0
networkx==3.1
python-dateutil==2.8.2




